Split up prototype into "micronaut-camunda-feature" and the example application "micronaut-camunda-app"
