--- micronaut-camunda-app/.gitignore
@@ -1,14 +1,0 @@
-Thumbs.db
-.DS_Store
-.gradle
-build/
-target/
-out/
-.idea
-*.iml
-*.ipr
-*.iws
-.project
-.settings
-.classpath
-dependency-reduced-pom.xml
