+++ micronaut-camunda-feature/gradle.properties
@@ -1,0 +1,1 @@
+micronautVersion=1.3.0
